More -Wconversion fixes.
git-svn-id: https://svn.r-project.org/R/trunk@55420 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/CommandLineArgs.c
@@ -56,7 +56,7 @@ R_set_command_line_arguments(int argc, char **argv)
     int i;
 
     NumCommandLineArgs = argc;
-    CommandLineArgs = (char**) calloc(argc, sizeof(char*));
+    CommandLineArgs = (char**) calloc((size_t) argc, sizeof(char*));
 
     for(i = 0; i < argc; i++)
 	CommandLineArgs[i] = strdup(argv[i]);
